feat(css): add css rule parser
