+++ src/lib.rs
@@ -2,7 +2,7 @@ extern crate alloc;
 
 pub mod api;
 #[allow(dead_code)]
-pub mod css3_parser;
+pub mod css;
 #[allow(dead_code)]
 pub mod html5_parser;
 pub mod testing;
